fix: remove doorbell press simulation options
BREAKING CHANGE: The `sendDoorbellMotionNotificationsToTv` and `sendCameraMotionNotificationsToTv` options have been removed.  These options were originally inteded as a workaround for the fact that tvOS would only automatically start streaming for doorbell presses and not for motion events.  tvOS 15 now supports this features as an option in the UI, so we no longer need this option at the plugin level.